refactor(scale_pattern): rename test file
