+++ test/src/scale/scale_pattern_test.dart
